requirements.txt
Flask==2.3.3
yfinance==0.2.61
pandas==2.1.1
numpy==1.24.3
APScheduler==3.10.4
requests==2.31.0
Werkzeug==2.3.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
curl_cffi
